PROD-1118: Update django dependency for django2.
