+++ Makefile
@@ -77,7 +77,7 @@ requirements: ## install development environment requirements
 	pip-sync requirements/dev.txt requirements/private.*
 
 test: clean ## run tests in the current virtualenv
-	pytest
+	PYTHONPATH=./:.. pytest
 
 test-all: ## run tests on every supported Python/Django combination
 	tox -e quality
