+++ manage.py
@@ -12,6 +12,7 @@ PWD = os.path.abspath(os.path.dirname(__file__))
 
 if __name__ == '__main__':
     os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'test_settings')
+    sys.path.append('./mock_apps')
     sys.path.append(PWD)
     try:
         from django.core.management import execute_from_command_line  # pylint: disable=wrong-import-position
