+++ mock_apps/example/apps.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+"""
+config_models Django application initialization.
+"""
+
+from __future__ import absolute_import, unicode_literals
+
+from django.apps import AppConfig
+
+
+class ExampleConfig(AppConfig):
+    """
+    Configuration for the config_models Django application.
+    """
+
+    name = 'example'
