+++ requirements/dev.txt
@@ -9,7 +9,6 @@ astroid==1.6.6
 atomicwrites==1.3.0
 attrs==19.3.0
 backports.functools-lru-cache==1.6.1
-backports.os==0.1.1       # via path.py
 caniusepython3==7.2.0
 certifi==2019.11.28
 chardet==3.0.4
@@ -18,7 +17,7 @@ click==7.0
 codecov==2.0.15
 configparser==4.0.2
 contextlib2==0.6.0.post1
-coverage==5.0.1
+coverage==5.0.2
 ddt==1.2.2
 distlib==0.3.0
 django==1.11.27
@@ -29,7 +28,6 @@ enum34==1.1.6
 filelock==3.0.12
 freezegun==0.3.12
 funcsigs==1.0.2
-future==0.18.2            # via backports.os
 futures==3.2.0 ; python_version == "2.7"
 idna==2.8
 importlib-metadata==1.3.0
@@ -38,7 +36,7 @@ lazy-object-proxy==1.4.3
 mccabe==0.6.1
 mock==3.0.5
 more-itertools==5.0.0
-packaging==19.2
+packaging==20.0
 path.py==11.5.2           # via edx-i18n-tools
 pathlib2==2.3.5
 pip-tools==4.3.0
@@ -54,10 +52,10 @@ pylint==1.9.5
 pyparsing==2.4.6
 pytest-cov==2.8.1
 pytest-django==3.7.0
-pytest==4.6.8
+pytest==4.6.9
 python-dateutil==2.8.1
 pytz==2019.3
-pyyaml==5.2               # via edx-i18n-tools
+pyyaml==5.3               # via edx-i18n-tools
 requests==2.22.0
 scandir==1.10.0
 singledispatch==3.4.0.3
@@ -66,10 +64,10 @@ snowballstemmer==2.0.0
 toml==0.10.0
 tox-battery==0.5.1
 tox-travis==0.12
-tox==3.14.2
+tox==3.14.3
 urllib3==1.25.7
 virtualenv==16.7.9
-wcwidth==0.1.7
+wcwidth==0.1.8
 wrapt==1.11.2
 zipp==0.6.0
