+++ requirements/doc.txt
@@ -7,13 +7,13 @@
 alabaster==0.7.12         # via sphinx
 atomicwrites==1.3.0
 attrs==19.3.0
-babel==2.7.0              # via sphinx
+babel==2.8.0              # via sphinx
 bleach==3.1.0             # via readme-renderer
 certifi==2019.11.28       # via requests
 chardet==3.0.4            # via requests
 configparser==4.0.2
 contextlib2==0.6.0.post1
-coverage==5.0.1
+coverage==5.0.2
 ddt==1.2.2
 django==1.11.27
 djangorestframework==3.9.4
@@ -27,7 +27,7 @@ jinja2==2.10.3            # via sphinx
 markupsafe==1.1.1         # via jinja2
 mock==3.0.5
 more-itertools==5.0.0
-packaging==19.2
+packaging==20.0
 pathlib2==2.3.5
 pkginfo==1.5.0.1          # via twine
 pluggy==0.13.1
@@ -36,7 +36,7 @@ pygments==2.5.2           # via readme-renderer, sphinx
 pyparsing==2.4.6
 pytest-cov==2.8.1
 pytest-django==3.7.0
-pytest==4.6.8
+pytest==4.6.9
 python-dateutil==2.8.1
 pytz==2019.3
 readme-renderer==24.0
@@ -48,11 +48,11 @@ snowballstemmer==2.0.0    # via sphinx
 sphinx-rtd-theme==0.4.3
 sphinx==1.8.5
 sphinxcontrib-websupport==1.1.2  # via sphinx
-tqdm==4.41.0              # via twine
+tqdm==4.41.1              # via twine
 twine==1.15.0
 typing==3.7.4.1           # via sphinx
 urllib3==1.25.7           # via requests
-wcwidth==0.1.7
+wcwidth==0.1.8
 webencodings==0.5.1       # via bleach
 zipp==0.6.0
