+++ requirements/quality.txt
@@ -16,7 +16,7 @@ click-log==0.3.2          # via edx-lint
 click==7.0                # via click-log, edx-lint
 configparser==4.0.2
 contextlib2==0.6.0.post1
-coverage==5.0.1
+coverage==5.0.2
 ddt==1.2.2
 distlib==0.3.0            # via caniusepython3
 django==1.11.27
@@ -33,7 +33,7 @@ lazy-object-proxy==1.4.3  # via astroid
 mccabe==0.6.1             # via pylint
 mock==3.0.5
 more-itertools==5.0.0
-packaging==19.2
+packaging==20.0
 pathlib2==2.3.5
 pluggy==0.13.1
 py==1.8.1
@@ -46,7 +46,7 @@ pylint==1.9.5             # via edx-lint, pylint-celery, pylint-django, pylint-p
 pyparsing==2.4.6
 pytest-cov==2.8.1
 pytest-django==3.7.0
-pytest==4.6.8
+pytest==4.6.9
 python-dateutil==2.8.1
 pytz==2019.3
 requests==2.22.0          # via caniusepython3
@@ -55,7 +55,7 @@ singledispatch==3.4.0.3   # via astroid, pylint
 six==1.13.0
 snowballstemmer==2.0.0    # via pydocstyle
 urllib3==1.25.7           # via requests
-wcwidth==0.1.7
+wcwidth==0.1.8
 wrapt==1.11.2             # via astroid
 zipp==0.6.0
