+++ requirements/test.txt
@@ -8,24 +8,24 @@ atomicwrites==1.3.0       # via pytest
 attrs==19.3.0             # via pytest
 configparser==4.0.2       # via importlib-metadata
 contextlib2==0.6.0.post1  # via importlib-metadata
-coverage==5.0.1           # via pytest-cov
+coverage==5.0.2           # via pytest-cov
 ddt==1.2.2
 freezegun==0.3.12
 funcsigs==1.0.2           # via mock, pytest
 importlib-metadata==1.3.0  # via pluggy, pytest
 mock==3.0.5
 more-itertools==5.0.0     # via pytest, zipp
-packaging==19.2           # via pytest
+packaging==20.0           # via pytest
 pathlib2==2.3.5           # via importlib-metadata, pytest, pytest-django
 pluggy==0.13.1            # via pytest
 py==1.8.1                 # via pytest
 pyparsing==2.4.6          # via packaging
 pytest-cov==2.8.1
 pytest-django==3.7.0
-pytest==4.6.8             # via pytest-cov, pytest-django
+pytest==4.6.9             # via pytest-cov, pytest-django
 python-dateutil==2.8.1    # via freezegun
 pytz==2019.3
 scandir==1.10.0           # via pathlib2
 six==1.13.0               # via freezegun, mock, packaging, pathlib2, pytest, python-dateutil
-wcwidth==0.1.7            # via pytest
+wcwidth==0.1.8            # via pytest
 zipp==0.6.0               # via importlib-metadata
