+++ requirements/travis.txt
@@ -9,12 +9,12 @@ chardet==3.0.4            # via requests
 codecov==2.0.15
 configparser==4.0.2       # via importlib-metadata
 contextlib2==0.6.0.post1  # via importlib-metadata
-coverage==5.0.1           # via codecov
+coverage==5.0.2           # via codecov
 filelock==3.0.12          # via tox
 idna==2.8                 # via requests
 importlib-metadata==1.3.0  # via pluggy, tox
 more-itertools==5.0.0     # via zipp
-packaging==19.2           # via tox
+packaging==20.0           # via tox
 pathlib2==2.3.5           # via importlib-metadata
 pluggy==0.13.1            # via tox
 py==1.8.1                 # via tox
@@ -25,7 +25,7 @@ six==1.13.0               # via packaging, pathlib2, tox
 toml==0.10.0              # via tox
 tox-battery==0.5.1
 tox-travis==0.12
-tox==3.14.2
+tox==3.14.3
 urllib3==1.25.7           # via requests
 virtualenv==16.7.9        # via tox
 zipp==0.6.0               # via importlib-metadata
