+++ tests/test_decorators.py
@@ -3,20 +3,10 @@ Tests for config models decorators
 """
 
 from __future__ import absolute_import, unicode_literals
-from django.utils import six
 
 from config_models import decorators
-from config_models.models import ConfigurationModel
-from config_models.tests.utils import CacheIsolationTestCase
-
-# pylint: disable=model-missing-unicode
-@six.python_2_unicode_compatible
-class ExampleDecoratorConfig(ConfigurationModel):
-    """
-    Test model for testing the require_config decorator
-    """
-    def __str__(self):
-        return "ExampleDecoratorConfig(enabled={})".format(self.enabled)
+from example.models import ExampleDecoratorConfig
+from tests.utils import CacheIsolationTestCase
 
 
 @decorators.require_config(ExampleDecoratorConfig)
