--- tests/test_models.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-"""
-Tests for the `config_models` models module.
-"""
-
-from __future__ import absolute_import, unicode_literals
