+++ tox.ini
@@ -1,7 +1,7 @@
 [tox]
-envlist = py{27,35}-django{111}-drf{37,38,39,latest},py{35,36}-django{20,21,22}-drf{37,38,39,latest}
+envlist = py{27,35}-django{111}-drf{39,latest},py{35,36}-django{20,21,22}-drf{39,latest}
 
-[pep8]
+[pycodestyle]
 exclude = .git,.tox,migrations
 max-line-length = 120
 
@@ -10,11 +10,13 @@ ignore = D200,D203
 match-dir = (?!migrations)
 
 [pytest]
-DJANGO_SETTINGS_MODULE = test_settings
 addopts = --cov config_models --cov-report term-missing
 norecursedirs = .* docs requirements
 
 [testenv]
+setenv =
+    PYTHONPATH = {toxinidir}/mock_apps
+    DJANGO_SETTINGS_MODULE = test_settings
 deps =
     django111: Django>=1.11,<2
     django20: Django>=2.0,<2.1
@@ -31,7 +33,7 @@ commands =
 [testenv:docs]
 setenv =
     DJANGO_SETTINGS_MODULE = test_settings
-    PYTHONPATH = {toxinidir}
+    PYTHONPATH = {toxinidir}/:{toxinidir}/mock_apps
 whitelist_externals =
     make
     rm
@@ -54,12 +56,10 @@ whitelist_externals =
 deps =
     -r{toxinidir}/requirements/quality.txt
 commands =
-    touch tests/__init__.py
     pylint config_models
     pylint tests
     pylint --py3k config_models
     pylint --py3k tests
-    rm tests/__init__.py
     pycodestyle config_models tests
     #pydocstyle config_models tests
     make help
